+++ Google-Cloud/Service-Accounts/gcp-verify/HISTORY.md
@@ -1,2 +1,3 @@
 # Program History
  - 2022-12-12 - released to repository
+ - 2022-12-13 - added Makefile.linux for building on Ubuntu